isolate PROGMEM
Isolated dependencies

sensors dependencies

enums, macros and prototypes.

I2C dependencies moved to actual sensor files.

definitelly redefine
