+++ code/espurna/config/defaults.h
@@ -404,14 +404,6 @@
 #define HOSTNAME ""
 #endif
 
-// Needed for ESP8285 boards under Windows using PlatformIO (?)
-#ifndef BUTTON_PUSHBUTTON
-#define BUTTON_PUSHBUTTON   0
-#define BUTTON_SWITCH       1
-#define BUTTON_DEFAULT_HIGH 2
-#define BUTTON_SET_PULLUP   4
-#endif
-
 // Relay providers
 #ifndef RELAY_PROVIDER
 #define RELAY_PROVIDER          RELAY_PROVIDER_RELAY
