+++ code/espurna/sensors/SI7021Sensor.h
@@ -7,6 +7,11 @@
 
 #pragma once
 
+#undef I2C_SUPPORT
+#define I2C_SUPPORT 1 // Explicitly request I2C support.
+
+
+
 #include "Arduino.h"
 #include "I2CSensor.h"
